+++ src/components/layout/side-bar/user-profile/index.tsx
@@ -1,6 +1,6 @@
 import { Button, Space } from 'antd';
 import styled from 'styled-components';
-import TextBox from '@components/text';
+import TextBox from '@components/text-box';
 import { colors } from '@/constants/colors';
 import { StyledSpaceProps } from './type';
