+++ src/components/text-box/type.ts
